requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
gunicorn==21.2.0
werkzeug==2.3.7
flask-cors==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0 